minor chip fix

+++ app/src/main/java/com/example/goodnote/ui/viewModels/NoteDetailsViewModel.kt
@@ -65,13 +65,19 @@ class NoteDetailsViewModel(private val noteRepo: NoteRepo,
     fun deleteTagForNote(noteId: String, tagId: String) = viewModelScope.launch {
         val updateTags: MutableList<TagModel> = noteToEdit.value?.tags?.filter { it.tagId != tagId }?.toMutableList() ?: mutableListOf()
         _noteToEdit.value = noteToEdit.value?.copy(tags=updateTags)
-        withContext(Dispatchers.IO) {noteRepo.deleteTagForNote(noteId, tagId)}
+        if (iddd.isNotBlank()) {
+            withContext(Dispatchers.IO) {noteRepo.deleteTagForNote(noteId, tagId)}
+        }
     }
 
     fun addTagForNote(noteId: String, tag: TagModel) = viewModelScope.launch {
+        val newTags = _noteToEdit.value.tags + listOf(tag)
         _noteToEdit.value?.tags?.add(tag)
         _noteToEdit.value = _noteToEdit.value?.copy()
-        withContext(Dispatchers.IO) {noteRepo.addTagForNote(noteId, tag.tagId)}
+
+        if (iddd.isNotBlank()) {
+            withContext(Dispatchers.IO) {noteRepo.addTagForNote(noteId, tag.tagId)}
+        }
     }
 
 }
